+++ Lab3/__init__.py
@@ -1,2 +1,2 @@
-from .json import JsonSerializer
-from .xml import XMLSerializer
+from .json_serializer import JsonSerializer
+from .xml_serializer import XMLSerializer
