+++ Lab3/serializer_factory.py
@@ -1,6 +1,5 @@
-from json import JsonSerializer
-from xml import XMLSerializer
-
+from json_serializer import JsonSerializer
+from  xml_serializer import XMLSerializer
 class SerializerFactory:
 
     @staticmethod
